requirements.txt
openai
pdfplumber
fastapi[standard]
uvicorn
python-dotenv
pypandoc
python-docx